# api/spec/openapi.yaml
openapi: 3.0.3
info:
  title: RaidReplay API
  description: Self-hosted WoW raid recording and replay server. Spec is also
    served live at GET /api/openapi.yaml.
  version: 0.1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/recordings/upload-url:
    post:
      summary: Request an upload URL for a new recording
      tags:
        - recordings
      description: The client calls this before uploading a video. Returns a
        recordingId and a URL to PUT the video file to. For S3 storage the URL
        points directly to S3 (presigned). For local storage the URL points to
        this API.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fightId
                - encounterID
                - encounterName
                - difficultyID
                - groupSize
                - kill
                - durationMs
                - videoStartUnixMs
                - pullNumber
              properties:
                fightId:
                  format: uuid
                  type: string
                encounterID:
                  type: number
                encounterName:
                  type: string
                difficultyID:
                  type: number
                groupSize:
                  minimum: 1
                  maximum: 40
                  type: number
                kill:
                  type: boolean
                bossPct:
                  minimum: 0
                  maximum: 100
                  type: number
                durationMs:
                  minimum: 0
                  type: number
                videoStartUnixMs:
                  type: number
                pullNumber:
                  minimum: 1
                  type: number
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - recordingId
                  - uploadUrl
                  - expiresAt
                properties:
                  recordingId:
                    format: uuid
                    type: string
                  uploadUrl:
                    format: uri
                    type: string
                  expiresAt:
                    format: date-time
                    type: string
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                properties:
                  error:
                    type: string
  /api/v1/recordings/{id}/upload:
    put:
      summary: Upload video file
      tags:
        - recordings
      description: PUT the raw video bytes here. Only used when storage backend is
        local disk. For S3, the client uploads directly to the presigned URL.
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - fightId
                  - userId
                  - status
                  - durationMs
                  - videoStartUnixMs
                  - fileSizeBytes
                  - format
                  - createdAt
                properties:
                  id:
                    format: uuid
                    type: string
                  fightId:
                    format: uuid
                    type: string
                  userId:
                    anyOf:
                      - format: uuid
                        type: string
                      - type: "null"
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - uploading
                      - type: string
                        enum:
                          - processing
                      - type: string
                        enum:
                          - ready
                      - type: string
                        enum:
                          - error
                  durationMs:
                    type: number
                  videoStartUnixMs:
                    type: number
                  fileSizeBytes:
                    anyOf:
                      - type: number
                      - type: "null"
                  format:
                    type: string
                  createdAt:
                    format: date-time
                    type: string
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                properties:
                  error:
                    type: string
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                properties:
                  error:
                    type: string
  /api/v1/recordings/{id}/stream:
    get:
      summary: Stream video file
      tags:
        - recordings
      description: For local storage, streams the file directly. For S3 storage,
        redirects to a presigned download URL.
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                properties:
                  error:
                    type: string
  /api/v1/recordings/{id}:
    get:
      summary: Get recording metadata
      tags:
        - recordings
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - fightId
                  - userId
                  - status
                  - durationMs
                  - videoStartUnixMs
                  - fileSizeBytes
                  - format
                  - createdAt
                properties:
                  id:
                    format: uuid
                    type: string
                  fightId:
                    format: uuid
                    type: string
                  userId:
                    anyOf:
                      - format: uuid
                        type: string
                      - type: "null"
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - uploading
                      - type: string
                        enum:
                          - processing
                      - type: string
                        enum:
                          - ready
                      - type: string
                        enum:
                          - error
                  durationMs:
                    type: number
                  videoStartUnixMs:
                    type: number
                  fileSizeBytes:
                    anyOf:
                      - type: number
                      - type: "null"
                  format:
                    type: string
                  createdAt:
                    format: date-time
                    type: string
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                properties:
                  error:
                    type: string
    delete:
      summary: Delete a recording
      tags:
        - recordings
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Default Response
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                properties:
                  error:
                    type: string
  /api/v1/recordings/{id}/status:
    patch:
      summary: Update recording status
      tags:
        - recordings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  anyOf:
                    - type: string
                      enum:
                        - uploading
                    - type: string
                      enum:
                        - processing
                    - type: string
                      enum:
                        - ready
                    - type: string
                      enum:
                        - error
                fileSizeBytes:
                  type: number
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - fightId
                  - userId
                  - status
                  - durationMs
                  - videoStartUnixMs
                  - fileSizeBytes
                  - format
                  - createdAt
                properties:
                  id:
                    format: uuid
                    type: string
                  fightId:
                    format: uuid
                    type: string
                  userId:
                    anyOf:
                      - format: uuid
                        type: string
                      - type: "null"
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - uploading
                      - type: string
                        enum:
                          - processing
                      - type: string
                        enum:
                          - ready
                      - type: string
                        enum:
                          - error
                  durationMs:
                    type: number
                  videoStartUnixMs:
                    type: number
                  fileSizeBytes:
                    anyOf:
                      - type: number
                      - type: "null"
                  format:
                    type: string
                  createdAt:
                    format: date-time
                    type: string
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                properties:
                  error:
                    type: string
...